pyproject.toml
[project]
name = "python-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "pillow>=10.4.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.3",
]
